Apply suggested changes

+++ ee/tabby-webserver/src/service/auth.rs
@@ -227,7 +227,7 @@ impl AuthenticationService for AuthenticationServiceImpl {
         };
 
         let id = user.id.as_rowid()?;
-        let existing = self.db.get_password_reset_by_user_id(id).await?;
+        let existing = self.db.get_password_reset_by_user_id(id as i64).await?;
         if let Some(existing) = existing {
             if Utc::now().signed_duration_since(&*existing.created_at) < Duration::minutes(5) {
                 return Err(anyhow!(
@@ -235,7 +235,7 @@ impl AuthenticationService for AuthenticationServiceImpl {
                 ));
             }
         }
-        let code = self.db.create_password_reset(id).await?;
+        let code = self.db.create_password_reset(id as i64).await?;
         self.mail
             .send_password_reset_email(user.email, code.clone())
             .await?;
@@ -247,7 +247,9 @@ impl AuthenticationService for AuthenticationServiceImpl {
             password_hash(password).map_err(|_| PasswordResetError::Unknown)?;
 
         let user_id = self.db.verify_password_reset(code).await?;
-        self.db.delete_password_reset_by_user_id(user_id).await?;
+        self.db
+            .delete_password_reset_by_user_id(user_id as i64)
+            .await?;
         self.db
             .update_user_password(user_id, password_encrypted)
             .await?;
@@ -908,7 +910,7 @@ mod tests {
             .unwrap();
         let reset = service
             .db
-            .get_password_reset_by_user_id(user.id.as_rowid().unwrap())
+            .get_password_reset_by_user_id(user.id.as_rowid().unwrap() as i64)
             .await
             .unwrap()
             .unwrap();
@@ -931,7 +933,7 @@ mod tests {
             .unwrap();
         let reset = service
             .db
-            .get_password_reset_by_user_id(user.id)
+            .get_password_reset_by_user_id(user.id as i64)
             .await
             .unwrap()
             .unwrap();
@@ -960,7 +962,7 @@ mod tests {
             .unwrap();
         let reset = service
             .db
-            .get_password_reset_by_user_id(user_id_2)
+            .get_password_reset_by_user_id(user_id_2 as i64)
             .await
             .unwrap()
             .unwrap();
